+++ frontend/ChangeLog.md
@@ -9,3 +9,10 @@ Moreover, we provide seven themes, you can select one that you like.
 For a smoother user experience, we delete the logic of "fileUploaded". 
 
 We use Ant Design for a more advanced and concise design. 
+
+# 0.1.2
+
+Use more Ant Design
+
+Add Electron to package project, can use "npm run" to build & package. 
+No longer run on the browser, but a separate window.